social/requirements.txt
discord.py
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
openai
tiktoken
facebook-sdk
boto3
pillow
